interview-backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
supabase==2.10.0
livekit>=1.0,<2.0
livekit-agents>=1.5,<2.0
livekit-plugins-groq>=1.5,<2.0
livekit-plugins-deepgram>=1.5,<2.0
livekit-plugins-azure>=1.5,<2.0
livekit-plugins-silero>=1.5,<2.0
aiohttp==3.11.0
python-dotenv==1.0.1
